refactor(tests): extract slot-mock rig into fixtures/slot_mocks.rs
FixedUpstream, TaggedCatalog, and the spawn_mock_upstream_with_slots /
spawn_proxy_with_cache* helpers were private to integration_slot_roundtrip.rs.
Pulling them into a dedicated fixture file lets a second test file reuse the
same mock llama-server + proxy-spawn rig instead of duplicating it.
