+++ crates/gglib-proxy/tests/fixtures/mod.rs
@@ -1,4 +1,5 @@
 //! Test fixtures shared by proxy integration tests.
 #![allow(dead_code)]
 pub mod common;
+pub mod slot_mocks;
 pub mod sse;
